+++ qbraid/visualization/__init__.py
@@ -25,7 +25,6 @@ Functions
     plot_conversion_graph
     plot_runtime_conversion_scheme
     circuit_drawer
-    qasm3_drawer
     plot_atomic_register
     animate_qpu_state
 
@@ -40,7 +39,6 @@ Exceptions
 """
 from .ahs import plot_atomic_register
 from .draw_circuit import circuit_drawer
-from .draw_qasm3 import qasm3_drawer
 from .exceptions import VisualizationError
 from .flair_animations import animate_qpu_state
 from .plot_conversions import plot_conversion_graph, plot_runtime_conversion_scheme
@@ -52,7 +50,6 @@ __all__ = [
     "plot_conversion_graph",
     "plot_runtime_conversion_scheme",
     "circuit_drawer",
-    "qasm3_drawer",
     "plot_atomic_register",
     "animate_qpu_state",
     "VisualizationError",